test: pin the dispatch branch; drop a redundant conversion
Assert looksDense is false in TestDifferentialLongPatternDual's setup so
the test cannot silently start passing through the density branch of the
looksDense || chainHeavy dispatch if pattern shapes or thresholds shift.
Also drop the redundant uint32 conversion of rootState in chainWalk
(rootState is already uint32; every other scan loop binds it directly).

Reported by coderabbitai on PR #14.

+++ trie.go
@@ -574,7 +574,7 @@ func (tr *Trie) chainWalk(w []byte, skips, counted int) (int, int, bool) {
 		return skips, counted, true
 	}
 	counted += len(w) - j
-	s := uint32(rootState)
+	s := rootState
 	for i := j; i < len(w); i++ {
 		if s == rootState && w[i] != c {
 			k := bytes.IndexByte(w[i:], c)
